app/requirements.txt
streamlit>=1.30.0
watchdog>=3.0.0
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pydub>=0.25.1
moviepy>=1.0.3
python-dotenv>=1.0.0
requests>=2.28.1
opencv-python>=4.7.0
websocket-client>=1.6.0
